refactor: env and errors
