+++ src/ts/errors.ts
@@ -1,6 +1,7 @@
 export { ApiError } from "./errors/ApiError";
 export { BaseError } from "./errors/BaseError";
-export { PanicError, panic } from "./errors/PanicError";
-export { TypeError, validateType } from "./errors/TypeError";
-export { UnknownError, wrapErr } from "./errors/UnknownError";
+export { PanicError } from "./errors/PanicError";
+export { TypeError } from "./errors/TypeError";
+export { UnknownError } from "./errors/UnknownError";
 export { Result } from "./errors/Result";
+export { panic, wrapErr } from "./errors/utils";
