+++ src/ts/errors/ApiError.ts
@@ -1,5 +1,5 @@
-import { wrapErr } from "./UnknownError";
 import { BaseError } from "./BaseError";
+import { wrapErr } from "./utils";
 
 export class ApiError extends BaseError {
   public static wrap(statusCode: number, err: unknown): ApiError {
