+++ src/ts/errors/UnknownError.ts
@@ -5,13 +5,3 @@ export class UnknownError extends BaseError {
     super("UnknownError", cause != null ? "error is wrapped" : "", cause);
   }
 }
-
-export function wrapErr(err: unknown): BaseError {
-  if (err instanceof BaseError) {
-    return err;
-  }
-  if (err instanceof Error) {
-    return new UnknownError(err);
-  }
-  return new UnknownError(new Error(`${err}`, { cause: err }));
-}
